Trait eventually_v0_5_0::aggregate::Repository

pub trait Repository<T, R>: Send + Sync where
    T: Aggregate,
    R: Root<T>, {
    type Error;
    fn get<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        id: &'life1 T::Id
    ) -> Pin<Box<dyn Future<Output = Result<R, Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn store<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        root: &'life1 mut R
    ) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;
}

A Repository is an object that allows to load and save an Aggregate Root from and to a persistent data store.

Associated Types

type Error

The error type that can be returned by the Repository implementation during loading or storing of an Aggregate Root.

Required methods

fn get<'life0, 'life1, 'async_trait>(
&'life0 self,
id: &'life1 T::Id
) -> Pin<Box<dyn Future<Output = Result<R, Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Loads an Aggregate Root instance from the data store, referenced by its unique identifier.

fn store<'life0, 'life1, 'async_trait>(
&'life0 self,
root: &'life1 mut R
) -> Pin<Box<dyn Future<Output = Result<(), Self::Error>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Stores a new version of an Aggregate Root instance to the data store.